+++ assets/src/components/resource-details.js
@@ -1,5 +1,7 @@
 import React from "react";
 import * as PropTypes from "prop-types";
+import DOMPurify from "dompurify";
+import parse from "html-react-parser";
 import LoadingSpinner from "./loading-spinner";
 import getResourceFacilities from "../util/resource-utils";
 import "./resource-details.scss";
@@ -17,6 +19,10 @@ function ResourceDetails({ setShowResourceDetails, resource }) {
   const hideResourceView = () => {
     setShowResourceDetails(null);
   };
+
+  const sanitizedDescription = resource.resourceDescription
+    ? parse(DOMPurify.sanitize(resource.resourceDescription, {}))
+    : "";
 
   const getFacilitiesList = () => {
     const facilities = getResourceFacilities(resource);
@@ -47,7 +53,7 @@ function ResourceDetails({ setShowResourceDetails, resource }) {
       {resource && (
         <div>
           <div className="resource-headline">
-            <span>{resource.displayName ?? resource.resourceName}</span>
+            <span>{resource.resourceDisplayName ?? resource.resourceName}</span>
             <button type="button" className="booking-btn-inv" onClick={hideResourceView}>
               Tilbage til listen
             </button>
@@ -55,7 +61,7 @@ function ResourceDetails({ setShowResourceDetails, resource }) {
           <div className="resource-details row">
             <div className="image-wrapper col-xs-12 col-md-4">
               <div className="image">
-                <img alt={resource.displayName ?? resource.resourceName} src={resource.resourceImage} />
+                <img alt={resource.resourceDisplayName ?? resource.resourceName} src={resource.resourceImage} />
               </div>
             </div>
             <div className="facilities col-xs-12 col-md-4">
@@ -82,7 +88,7 @@ function ResourceDetails({ setShowResourceDetails, resource }) {
             <div className="resource-description">
               <span>Beskrivelse</span>
               <div>
-                <span>{resource.resourceDescription}</span>
+                <span>{sanitizedDescription}</span>
               </div>
             </div>
           )}
@@ -96,7 +102,7 @@ ResourceDetails.propTypes = {
   setShowResourceDetails: PropTypes.func.isRequired,
   resource: PropTypes.shape({
     capacity: PropTypes.number.isRequired,
-    displayName: PropTypes.string.isRequired,
+    resourceDisplayName: PropTypes.string.isRequired,
     resourceName: PropTypes.string.isRequired,
     location: PropTypes.string.isRequired,
     locationDisplayName: PropTypes.string,