+++ assets/src/components/user-booking-delete.js
@@ -63,7 +63,7 @@ function UserBookingDelete({ config, booking, onBookingDeleted, close }) {
               <div style={{ margin: "1em 0" }}>
                 <div>
                   <strong>Resource: </strong>
-                  {booking.displayName}
+                  {booking.resourceDisplayName}
                 </div>
                 <div>
                   <strong>Titel på booking: </strong>
@@ -99,7 +99,7 @@ UserBookingDelete.propTypes = {
   }).isRequired,
   booking: PropTypes.shape({
     id: PropTypes.string.isRequired,
-    displayName: PropTypes.string.isRequired,
+    resourceDisplayName: PropTypes.string.isRequired,
     subject: PropTypes.string.isRequired,
     start: PropTypes.string.isRequired,
     end: PropTypes.string.isRequired,