+++ CHANGELOG.md
@@ -8,7 +8,17 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ### Added
 
+- Use resourceDisplayName when displaying resources to the user.
+- Remove html from resourceDescription
+- Change map to use locationDisplayName for filter label and locationName as filter value.
+- Sorting to user panel.
+- Added text filtering to user panel.
+
 ### Changed
+
+- Added allowConflict option for allowing event overlap [#116](https://github.com/itk-dev/drupal_webform_booking_module/pull/116).
+- Removed NFS volume from docker setup [#116](https://github.com/itk-dev/drupal_webform_booking_module/pull/116).
+- Optimized calendar rendering [#116](https://github.com/itk-dev/drupal_webform_booking_module/pull/116).
 
 ## [1.0.1] - 2023-04-20